small update

+++ src/base-files/std-mem-man.cpp
@@ -1,4 +1,4 @@
-void reallocCharPtr(char*& src, long long new_size) {
+void reallocCharPtr(char*& src, int new_size) {
 	char* new_src = new char[new_size];
 
 	for(long long i = 0 ; i < new_size; ++i) {
