+++ src/base-files/std-mem-man.hpp
@@ -1,6 +1,6 @@
 #ifndef MEM_MAN
 #define MEM_MAN
 
-void reallocCharPtr(char*& src, long long new_len);
+void reallocCharPtr(char*& src, int new_len);
 
 #endif
